examples/recipe_sender/requirements.txt
beautifulsoup4==4.9.3
lxml==4.6.2
pandas==1.2.1
PyYAML==5.4.1
regex==2020.11.13
requests==2.25.1